+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.3.0 -> 1.0.0 (2015-10-06 <-> 2015-10-07)
+
+- Add readme description
+- Add rules to process enable and activate in the same method (1.0.0)
+- Add a method to enable a specific schema (0.9.0)
+- Add a method to add custom schema (0.8.0)
+- Add access method (0.7.0) - to enable passportJs / express / mongoose config
+- Add module passportJs (0.5.0) - default config for passportjs
+- Add module express  (0.4.0) -  default config for express
+- Add module mongoose  (0.3.0) -  default config for mongoose
+
 ## 0.2.1 (2015-10-06)
 
 - Hint program with yocto-hitn and fix bugs